api/requirements-pdf.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
google-generativeai==0.3.2
cohere==4.37
pydantic==2.4.2
qdrant-client==1.6.3
Pillow==10.0.1
pymupdf==1.23.5
